feat: 수정/삭제 시 서비스 계층에서 권한 검증 추가 (#14)
- 수정의 경우
  - 비회원이 작성한 글은 수정 불가
  - 회원이 작성한 글은 작성자 정보가 일치하는 경우에만 수정 가능
- 삭제의 경우
  - 비회원이 작성한 글은 모두 삭제 가능
  - 회원이 작성한 글은 작성자 정보가 일치하는 경우에만 수정 가능

* getById와 findById의 차이에 대해 정리하기!
- findById: 실제 값이 필요한 경우
- getById: 식별자 값만 필요한 경우 (프록시에 접근 - 지연 로딩)

+++ src/main/java/hyeong/lee/myboard/service/BoardService.java
@@ -1,11 +1,14 @@
 package hyeong.lee.myboard.service;
 
 import hyeong.lee.myboard.domain.Board;
+import hyeong.lee.myboard.domain.UserAccount;
 import hyeong.lee.myboard.dto.BoardRequestDto;
 import hyeong.lee.myboard.dto.BoardResponseDto;
 import hyeong.lee.myboard.dto.BoardWithRepliesResponseDto;
 import hyeong.lee.myboard.repository.BoardRepository;
+import hyeong.lee.myboard.repository.UserAccountRepository;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -13,12 +16,14 @@ import org.springframework.transaction.annotation.Transactional;
 
 import javax.persistence.EntityNotFoundException;
 
+@Slf4j
 @RequiredArgsConstructor
 @Transactional
 @Service
 public class BoardService {
 
     private final BoardRepository boardRepository;
+    private final UserAccountRepository userAccountRepository;
 
     @Transactional(readOnly = true) // 단건 읽기
     public BoardResponseDto readById(Long boardId) {
@@ -58,13 +63,38 @@ public class BoardService {
     }
 
     public void update(Long boardId, BoardRequestDto dto) {
+        // 실제 값이 필요하므로 findById
         Board board = findById(boardId);
-        board.updateContent(dto.getTitle(), dto.getContent());
+        // 식별자(userId)에 대한 값만 필요하므로 getById
+        UserAccount findUser = userAccountRepository.getReferenceById(dto.getUserAccountDto().getUserId());
+
+        if(board.getUserAccount() == null) {
+            log.error("비회원이 작성한 글은 수정할 수 없습니다"); return;
+        }
+
+        // 작성자가 일치하는 경우에만 수정 가능
+        if(board.getUserAccount().equals(findUser)) {
+            board.updateContent(dto.getTitle(), dto.getContent());
+        } else {
+            log.error(">> 사용자 정보가 일치하지 않습니다");
+        }
     }
 
-    public void delete(Long boardId) {
+    public void delete(Long boardId, UserAccount userAccount) {
         Board board = findById(boardId);
-        boardRepository.delete(board);
+
+        // 익명 회원이 작성한 글은 일단 모두 삭제 가능
+        if(board.getUserAccount() == null) {
+            boardRepository.delete(board);
+            return;
+        }
+
+        // 작성자 정보와 일치하는 경우에만 삭제 가능
+        if(board.getUserAccount().equals(userAccount)) {
+            boardRepository.delete(board);
+        } else {
+            log.error(">> 사용자 정보가 일치하지 않습니다");
+        }
     }
 
     @Transactional(readOnly = true)
